+++ classes/ui/wpdk-ui-controls.php
@@ -1078,9 +1078,11 @@ class WPDKUIControlCustom extends WPDKUIControl {
  *
  * @class              WPDKUIControlDate
  * @author             =undo= <info@wpxtre.me>
- * @copyright          Copyright (C) 2012-2013 wpXtreme Inc. All Rights Reserved.
- * @date               2012-11-28
- * @version            0.8.1
+ * @copyright          Copyright (C) 2012-2014 wpXtreme Inc. All Rights Reserved.
+ * @date               2014-09-20
+ * @version            0.9.0
+ *
+ * @history            0.9.0 - Create surrogate to fix date time string in timestamp
  *
  */
 class WPDKUIControlDate extends WPDKUIControl {
@@ -1107,7 +1109,30 @@ class WPDKUIControlDate extends WPDKUIControl {
    */
   public function draw()
   {
+    // Create surrogate - get original id and name
+    $original_id   = $this->id;
+    $original_name = $this->name;
+
+    // Build surrogate
+    $this->id   = sprintf( '%s-surrogate', $original_id );
+    $this->name = sprintf( '%s-surrogate', $original_name );
+
+    // Sanitize data
+    $this->item['data'] = isset( $this->item['data'] ) ? $this->item['data'] : array();
+
+    // Added reference
+    $this->item['data']['surrogate_name'] = $original_name;
+
     $this->inputType( WPDKHTMLTagInputType::TEXT, 'wpdk-form-date wpdk-form-has-button-clear-left ' );
+
+    // Create the input hidden
+    ?>
+    <input type="hidden"
+           data-type="wpdk-ui-date"
+           id="<?php echo $original_id ?>"
+           value="<?php echo empty( $this->item['value'] ) ? '' : strtotime( $this->item['value'] ) ?>"
+           name="<?php echo $original_name ?>" />
+    <?php
   }
 
 }
@@ -1134,9 +1159,11 @@ class WPDKUIControlDate extends WPDKUIControl {
  *
  * @class              WPDKUIControlDateTime
  * @author             =undo= <info@wpxtre.me>
- * @copyright          Copyright (C) 2012-2013 wpXtreme Inc. All Rights Reserved.
- * @date               2012-11-28
- * @version            0.8.1
+ * @copyright          Copyright (C) 2012-2014 wpXtreme Inc. All Rights Reserved.
+ * @date               2014-09-19
+ * @version            0.9.0
+ *
+ * @history            0.9.0 - Create surrogate to fix date time string in timestamp
  *
  */
 class WPDKUIControlDateTime extends WPDKUIControl {
@@ -1163,7 +1190,30 @@ class WPDKUIControlDateTime extends WPDKUIControl {
    */
   public function draw()
   {
+    // Create surrogate - get original id and name
+    $original_id   = $this->id;
+    $original_name = $this->name;
+
+    // Build surrogate
+    $this->id   = sprintf( '%s-surrogate', $original_id );
+    $this->name = sprintf( '%s-surrogate', $original_name );
+
+    // Sanitize data
+    $this->item['data'] = isset( $this->item['data'] ) ? $this->item['data'] : array();
+
+    // Added reference
+    $this->item['data']['surrogate_name'] = $original_name;
+
     $this->inputType( WPDKHTMLTagInputType::TEXT, 'wpdk-form-datetime wpdk-form-has-button-clear-left ' );
+
+    // Create the input hidden
+    ?>
+    <input type="hidden"
+           data-type="wpdk-ui-date-time"
+           id="<?php echo $original_id ?>"
+           value="<?php echo empty( $this->item['value'] ) ? '' : strtotime( $this->item['value'] ) ?>"
+           name="<?php echo $original_name ?>" />
+    <?php
   }
 
 }