+++ CHANGELOG.md
@@ -20,8 +20,14 @@ For more information on SemVer, please visit http://semver.org/.
 
 ---
 
-## Version 1.5.16
+## Version 1.5.17
 ### 2014-09-??
+
+* Several improved and fixed on list table view controller
+* Improved action filter and post data processing in list table model and view controller
+
+## Version 1.5.16
+### 2014-09-19
 
 #### Improvements